requirements.txt
streamlit
pymupdf
google.generativeai
regex